Modified 6-size.c

+++ 0x00-hello_world/6-size.c
@@ -14,11 +14,11 @@ int main(void)
 	long long int l;
 	float m;
 
-	printf("Size of char: %lu bytes(s)\n", (unsigned long)sizeof(i));
-	printf("Size of int: %lu bytes(s)\n", (unsigned long)sizeof(j));
-	printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(k));
-	printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(l));
-	printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(m));
+	printf("Size of char: %lu bytes(s)\n", sizeof(i));
+	printf("Size of int: %lu bytes(s)\n", sizeof(j));
+	printf("Size of long int: %lu byte(s)\n", sizeof(k));
+	printf("Size of long long int: %lu byte(s)\n", sizeof(l));
+	printf("Size of float: %lu byte(s)\n", sizeof(m));
 
 	return (0);
 }
